Add batchSize, flushInterval to topic struct and remove useless
