+++ coordinator-helper/helper/topic_managing_test.go
@@ -20,6 +20,46 @@ func newTopicManagingTestClient() *topicManagingTestClient {
 	}
 }
 
+func TestMakeTopicData(t *testing.T) {
+	var expectedDescription = "test topic meta"
+	var expectedNumFragments uint32 = 1
+	var expectedReplicationFactor uint32 = 0
+	var expectedNumPublishers uint64 = 1
+
+	data := NewFrameForTopicFromValues(expectedDescription, expectedNumFragments, expectedReplicationFactor, expectedNumPublishers)
+
+	if data.Description() != expectedDescription {
+		t.Error("Description is not equal")
+	}
+
+	if data.NumFragments() != expectedNumFragments {
+		t.Error("Num Fragment is not equal")
+	}
+
+	if data.ReplicationFactor() != expectedReplicationFactor {
+		t.Error("Replication Factor is not equal")
+	}
+
+	if data.NumPublishers() != expectedNumPublishers {
+		t.Error("NumPublisher is not equal")
+	}
+}
+
+func TestMakeTopicFragmentData(t *testing.T) {
+	var expectedLastOffset uint64 = 1
+	var expectedNumSubscribers uint64 = 1
+
+	data := NewFrameForFragmentFromValues(expectedLastOffset, expectedNumSubscribers)
+
+	if data.LastOffset() != expectedLastOffset {
+		t.Error("LastOffset is not equal")
+	}
+
+	if data.NumSubscribers() != expectedNumSubscribers {
+		t.Error("NumSubscribers is not equal")
+	}
+}
+
 func TestTopicManagingHelper_AddTopic(t *testing.T) {
 	testClient := newTopicManagingTestClient()
 	expectedTopic := "topic1"
