+++ proto/data.proto
@@ -28,4 +28,6 @@ message Topic {
     }
     string topic_name = 1;
     repeated FragmentOffset offsets = 2;
+    uint32 max_batch_size = 3;
+    uint32 flush_interval = 4;
 }
